modified the GPU kernel in order to support CUDA 6.0 or higher

+++ CAR_TRACKING/GPU_function.cu
@@ -1505,7 +1505,7 @@ calc_hist
       FLOAT vy0 = yp - iyp;
       FLOAT vx1 = 1.0 - vx0;
       FLOAT vy1 = 1.0 - vy0;
-      v = sqrt((double)v);
+      v = sqrtf((double)v);
       
       
 #ifdef USE_FLOAT_AS_DECIMAL
@@ -1776,22 +1776,22 @@ calc_feat
       //      p = norm + (x+1)*blocks[0] + y+1;
       p = norm + (x+1)*blocks_0 + y+1;
       //      n1 = 1.0 / sqrt(*p + *(p+1) + *(p+blocks[0]) + *(p+blocks[0]+1) + EPS);
-      n1 = 1.0 / sqrt(*p + *(p+1) + *(p+blocks_0) + *(p+blocks_0+1) + EPS);
+      n1 = 1.0 / sqrtf(*p + *(p+1) + *(p+blocks_0) + *(p+blocks_0+1) + EPS);
       
       //      p = norm + (x+1)*blocks[0] + y;
       p = norm + (x+1)*blocks_0 + y;
       //      n2 = 1.0 / sqrt(*p + *(p+1) + *(p+blocks[0]) + *(p+blocks[0]+1) + EPS);
-      n2 = 1.0 / sqrt(*p + *(p+1) + *(p+blocks_0) + *(p+blocks_0+1) + EPS);
+      n2 = 1.0 / sqrtf(*p + *(p+1) + *(p+blocks_0) + *(p+blocks_0+1) + EPS);
       
       //      p = norm + x*blocks[0] + y+1;
       p = norm + x*blocks_0 + y+1;
       //      n3 = 1.0 / sqrt(*p + *(p+1) + *(p+blocks[0]) + *(p+blocks[0]+1) + EPS);
-      n3 = 1.0 / sqrt(*p + *(p+1) + *(p+blocks_0) + *(p+blocks_0+1) + EPS);
+      n3 = 1.0 / sqrtf(*p + *(p+1) + *(p+blocks_0) + *(p+blocks_0+1) + EPS);
       
       //      p = norm + x*blocks[0] + y;
       p = norm + x*blocks_0 + y;
       //      n4 = 1.0 / sqrt(*p + *(p+1) + *(p+blocks[0]) + *(p+blocks[0]+1) + EPS);
-      n4 = 1.0 / sqrt(*p + *(p+1) + *(p+blocks_0) + *(p+blocks_0+1) + EPS);
+      n4 = 1.0 / sqrtf(*p + *(p+1) + *(p+blocks_0) + *(p+blocks_0+1) + EPS);
 
       FLOAT t1 = 0;
       FLOAT t2 = 0;
